@@ -1,7 +1,7 @@
 class User < ApplicationRecord
   # Include default devise modules. Others available are:
-  # :confirmable, :lockable, :timeoutable, :trackable and :omniauthable
-  devise :database_authenticatable, :registerable,
+  #  :lockable, :timeoutable, :trackable and :omniauthable
+  devise :database_authenticatable, :confirmable, :registerable,
          :recoverable, :rememberable, :validatable
 
   
